Add Help mechanism + Rubocop fixes + minor optimizations
Help functionality allows user generate documentation for
artifacts right from the code. Essentially, the code already
contains a lot of information about artifact parameters, their
names, constraints and transformations. Help processor extracts
the information and generates documentation based on it

We has to refactor a few parts to improve introspectability of the
objects. In particular, Defaults module experienced heavy
modifications. They were made to have backwards-compatible usage
semantics. Namely, Definition set of classes was introduced to
enclose information about Defaults and to provide merge semantics.

Also, to split the functionality used by both help generators and
renders, abstract "Processor" entity was created. It is consumed
by both Render and Help. Also, there is a split in terminology:
- Render has a View
- Help generator has a Provider
- Processor has a Handler
Both View and Provider inherit from Processor::Handler.
Essentially both Render and Help are kinds of Processor.

Currently supported documentation generator by default is Rdoc.
It uses Rdoc objects and formatters to generate documentation.
It is able to generate documentation blocks in ASCII or HTML.
There is a plan to move out of Rdoc to something more suirable
for the purpose of programmatic text generation.

Additional batch of changes is due to strictening of Rubocop
rules in recent update of the version used by the library.

Optimizations:
- Introduce small utility to Module class so it is possible to get
  a collection of ancestors with specified parent
- Optimize code with the introduced Module functionality
- Small changes over the code to provide better introspection for
  Help generation