Refactor exceptions and DSL modules and bug fixes
= Bugfixes =

Inheritable module was silently accumulating all methods
across the entire project because ClassScope was a singleton.
Refactored and fixed as a result.

= Reasons for refactoring =

== Preparation for expression module integration ==

Concepts were added. Concept is a bundle consisting of
transformation and constraint. The name is used with the same
meaning as in C++ in par with "type classes" in Haskell and
"contracts" in Eiffel. You can chain concepts to get increased
specialization.

Expressions are syntactically and semantically checked clauses.
Users will be able to use either Exoressions Ruby DSL based on
Ruby features or string-based clauses which will be parsed
before semantics check. The reason is to avoid using evals and
provide compile-time checking when clauses are specified in Ruby
code with possibility to pass them as strings in structured
format like JSON as parameter block. Expressions are generally
useful for defining alarm triggers and metric math. So,
Expressions module is made to unify all use cases with the same
syntax.

== Refactoring of exception handling logic ==

The major change is that exception causes are used to carry
additional information and clarifications with separation of
exception formatting concern out of exception code. Before the
change, all exception messages were collated into a single
exception message blurb inside of the exception classes themselves.
Now, exception classes have strictly defined roles and carry only
the pieces of information related to these roles. Parameter
exceptions carry information about parameter and instance where the
issue appeared. Constraint exceptions which are wrapped by
parameter exceptions are carrying information about constraint
object and it's location and so on. To get full error message,
calling code is supposed to walk itself through the cause chain and
format the messages as needed.

== DSL elements ==

Before the refactoring, DSL definitions were made through DSL
modules which were containing factory methods for all DSL elements.
The problem were when you wanted to add new constraint,
transformation or getter, you would be required to create another
factory method inside of the DSL blocks. Now, you just need to
specify DSL element class which already has factory method name
defined inside of it. Single place to edit things.