Add expressions DSL
Expressions facility is a simple universal expression parser
which allows you to use the same syntax for a variery of use
cases. One of such cases is when you need to specify an expression
for monitor's trigger condition through different services
(CloudWatch for example) and agents (collectd).

Expressions provide both parser facility for the conditions
expressed as strings and Ruby DSL for the conditions expressed
through using domain language (Ruby for the moment). Having both
gives you an opportunity to specify different constraints,
thresholds and mathematical and logical formulas in general in a
universal way which is easy to find in code.

Also, Expressions can be put under constraints which gives user
the ability to semantically constrain what operations and what
operands are supported. For this, it's appropriate to use
"concepts" to bundle together constraint and Expression
transformation. All standard constraints are applicable to
expressions and their parameter, but the most useful ones are
is? and has?

All comparisons, logical operations and basic arithmetic
expressions are supported by the module (both parser and object
model). With the mentioned ability to apply constraints you can
limit semantics.