Refactor rendering and processing subsystems
In vast majority of cases you need to carry global mutable context
of the rendering/processing being performed.

Before this patch the framework was trying not to be opinionated
about the way users design view layer.

Now, there is only one option which is to create a subclass
of Processor/Render. If no mutable context needs to be maintained,
it is suggested to use singletons. Without context they are safe
to use even in multi-threaded environments.